requirements.txt
Django==5.1.1
channels==4.0.0
daphne==4.0.0
Pillow==10.4.0
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
channels-redis==4.2.0
dj-database-url==2.1.0
django-allauth==65.9.0
requests==2.32.4
PyJWT==2.10.1
